Move OSMTileDataJson out to GeoJsonFeatureCollection
This improves consistency with naming of GeoJsonGeometry and GeoJsonFeature, as well as reducing clutter.
